+++ Core/Dehacked/DehackedApplier.cs
@@ -961,6 +961,11 @@ public class DehackedApplier
         // This wasn't an option in the original game, needs to be applied to all potential projectile spawns.
         definition.Flags.SetRandomizeProjectile();
         definition.Properties.Height = 0;
+        definition.Properties.Radius = 0;
+        definition.Properties.Health = 0;
+        definition.Properties.Mass = 0;
+        definition.Properties.MonsterMovementSpeed = 0;
+        definition.Properties.MissileMovementSpeed = 0;
         composer.Add(definition);
         m_dehacked.DefinitionLookup[index] = definition;
         return definition;