+++ Tests/Unit/GameAction/Dehacked/DefaultFlags.cs
@@ -29,6 +29,19 @@ public class DefaultFlags
         newDef2.Flags.RandomizeProjectile().Should().BeTrue();
     }
 
+    [Fact(DisplayName = "New Dehacked thing has default properties")]
+    public void NewDehackedThingDefaultProperties()
+    {
+        var newDef = World.EntityManager.DefinitionComposer.GetByName("*deh/entity192");
+        newDef.Should().NotBeNull();
+        newDef.Properties.Height.Should().Be(0);
+        newDef.Properties.Radius.Should().Be(0);
+        newDef.Properties.Health.Should().Be(0);
+        newDef.Properties.Mass.Should().Be(0);
+        newDef.Properties.MissileMovementSpeed.Should().Be(0);
+        newDef.Properties.MonsterMovementSpeed.Should().Be(0);
+    }
+
     private static readonly string Dehacked =
 @"
 Thing 191 (NewThing)
@@ -38,5 +51,8 @@ Initial frame = 42069
 Thing 192 (NewThing with height)
 ID # = 192
 Height = 4521984
-Initial frame = 42069"";";
+Initial frame = 42069"";
+
+Thing 193 (DefaultThing)
+ID # = 193";
 }